Re-use consent even if access token expired
During authorization, if the `approval_prompt` parameter or the
`REQUEST_APPROVAL_PROMPT` setting is set to 'auto', a previous consent
is only considered if the access token is still valid. This severely
limits the timeframe where the approval can be skipped in practice.

This change considers all access tokens including expired ones. Expired
access tokens are kept around until the associated refresh token
expires or is revoked.